# docs/migration/openapi-step5.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaQuant API",
    "description": "迁移阶段本机接口；支持持久化后台任务和同步兼容接口。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_tasks_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BacktestTask"
                  },
                  {
                    "$ref": "#/components/schemas/OptimizationTask"
                  },
                  {
                    "$ref": "#/components/schemas/WalkForwardTask"
                  },
                  {
                    "$ref": "#/components/schemas/FactorTask"
                  },
                  {
                    "$ref": "#/components/schemas/DataTask"
                  },
                  {
                    "$ref": "#/components/schemas/AITask"
                  },
                  {
                    "$ref": "#/components/schemas/NLTask"
                  },
                  {
                    "$ref": "#/components/schemas/CombinationTask"
                  },
                  {
                    "$ref": "#/components/schemas/XTickTask"
                  },
                  {
                    "$ref": "#/components/schemas/ChatTask"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "backtest": "#/components/schemas/BacktestTask",
                    "optimization": "#/components/schemas/OptimizationTask",
                    "walk_forward": "#/components/schemas/WalkForwardTask",
                    "factor_evaluation": "#/components/schemas/FactorTask",
                    "data_update": "#/components/schemas/DataTask",
                    "ai_analysis": "#/components/schemas/AITask",
                    "nl_strategy": "#/components/schemas/NLTask",
                    "factor_combination": "#/components/schemas/CombinationTask",
                    "xtick_query": "#/components/schemas/XTickTask",
                    "ai_chat": "#/components/schemas/ChatTask"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Tasks",
        "operationId": "tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "QUEUED",
                    "RUNNING",
                    "CANCEL_REQUESTED",
                    "SUCCESS",
                    "PARTIAL",
                    "FAILED",
                    "CANCELLED",
                    "INTERRUPTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "backtest",
                    "optimization",
                    "walk_forward",
                    "factor_evaluation",
                    "data_update",
                    "ai_analysis",
                    "nl_strategy",
                    "factor_combination",
                    "xtick_query",
                    "ai_chat"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{identifier}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Task",
        "operationId": "task_api_v1_tasks__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factors/combinations/prepare": {
      "post": {
        "tags": [
          "advanced research"
        ],
        "summary": "Prepare Combination",
        "operationId": "prepare_combination_api_v1_factors_combinations_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CombinationPrepareInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AITask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "ai_analysis",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/AIInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "AITask"
      },
...
      "BacktestTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "backtest",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/RunInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "BacktestTask"
      },
...
      "ChatTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "ai_chat",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/ChatInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "ChatTask"
      },
...
      "CombinationPrepareInput": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "base_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "CombinationPrepareInput"
      },
      "CombinationTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "factor_combination",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/CombinationInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "CombinationTask"
      },
...
      "DataTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "data_update",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/UpdateInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "DataTask"
      },
...
      "FactorEvaluationInput": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "factor_name": {
            "type": "string",
            "title": "Factor Name"
          },
          "symbols": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbols"
          },
          "horizon": {
            "type": "integer",
            "enum": [
              1,
              5,
              10,
              20
            ],
            "title": "Horizon",
            "default": 5
          },
          "n_groups": {
            "type": "integer",
            "enum": [
              5,
              10
            ],
            "title": "N Groups",
            "default": 5
          },
          "neutralization": {
            "type": "string",
            "enum": [
              "none",
              "industry",
              "size",
              "both"
            ],
            "title": "Neutralization",
            "default": "none"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "factor_name"
        ],
        "title": "FactorEvaluationInput"
      },
      "FactorTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "factor_evaluation",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/FactorEvaluationInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "FactorTask"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NLTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "nl_strategy",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/NLInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "NLTask"
      },
      "OptimizationInput": {
        "properties": {
          "baseline_run_id": {
            "type": "string",
            "title": "Baseline Run Id"
          },
          "parameter_grid": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Parameter Grid"
          },
          "objective": {
            "type": "string",
            "enum": [
              "sharpe",
              "annual_return",
              "calmar",
              "max_drawdown"
            ],
            "title": "Objective",
            "default": "sharpe"
          },
          "max_drawdown_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown Limit"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "max_workers": {
            "type": "integer",
            "enum": [
              1,
              2,
              4
            ],
            "title": "Max Workers",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "baseline_run_id",
          "parameter_grid"
        ],
        "title": "OptimizationInput"
      },
      "OptimizationTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "optimization",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/OptimizationInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "OptimizationTask"
      },
...
      "RunInput": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/BacktestInput"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          },
          "plan_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Title"
          },
          "plan_inputs": {
            "type": "object",
            "title": "Plan Inputs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "RunInput"
      },
...
      "TaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "TaskListResponse"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "backtest",
              "optimization",
              "walk_forward",
              "factor_evaluation",
              "data_update",
              "ai_analysis",
              "nl_strategy",
              "factor_combination",
              "xtick_query",
              "ai_chat"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "CANCEL_REQUESTED",
              "SUCCESS",
              "PARTIAL",
              "FAILED",
              "CANCELLED",
              "INTERRUPTED"
            ],
            "title": "Status"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressResponse"
          },
          "error": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result_available": {
            "type": "boolean",
            "title": "Result Available"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "retry_of",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at",
          "cancel_requested",
          "progress",
          "error",
          "result_available"
        ],
        "title": "TaskResponse"
      },
...
      "WalkForwardInput": {
        "properties": {
          "baseline_run_id": {
            "type": "string",
            "title": "Baseline Run Id"
          },
          "parameter_grid": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Parameter Grid"
          },
          "objective": {
            "type": "string",
            "enum": [
              "sharpe",
              "annual_return",
              "calmar",
              "max_drawdown"
            ],
            "title": "Objective",
            "default": "sharpe"
          },
          "max_drawdown_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown Limit"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "training_months": {
            "type": "integer",
            "minimum": 3.0,
            "title": "Training Months",
            "default": 12
          },
          "test_months": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Test Months",
            "default": 3
          },
          "step_months": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Step Months",
            "default": 3
          },
          "max_windows": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Max Windows",
            "default": 12
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "baseline_run_id",
          "parameter_grid"
        ],
        "title": "WalkForwardInput"
      },
      "WalkForwardTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "walk_forward",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/WalkForwardInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "WalkForwardTask"
      },
...
      "XTickTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "xtick_query",
            "title": "Kind"
          },
          "input": {
            "$ref": "#/components/schemas/XTickInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "input"
        ],
        "title": "XTickTask"
      }
    }
  }
}